+++ SwiftCraftLauncher/GameFeature/Services/Java/JavaRuntimeDownloader.swift
@@ -60,7 +60,7 @@ class JavaRuntimeDownloader: @unchecked Sendable {
             .reduce(0, +)
 
         let semaphore = AsyncSemaphore(
-            value: DIContainer.shared.ui.generalSettingsManager.concurrentDownloads,
+            value: DIContainer.shared.ui.gameSettingsManager.concurrentDownloads,
         )
 
         let counter = AtomicCounter()
