+++ SwiftCraftLauncher/Resources/Localizable.xcstrings
@@ -10,9 +10,6 @@
     "-" : {
       "shouldTranslate" : false
     },
-    ":" : {
-      "shouldTranslate" : false
-    },
     "…" : {
       "shouldTranslate" : false
     },
